Conflicts:
	meta-digi-arm/recipes-bsp/u-boot/u-boot-dey_2015.04.bb
	meta-digi-arm/recipes-digi/trustfence/trustfence-sign-tools_git.bb
	meta-digi-arm/recipes-kernel/kernel-module-qualcomm/kernel-module-qualcomm.bb
	meta-digi-arm/recipes-kernel/linux/linux-dey.inc
	meta-digi-arm/recipes-kernel/linux/linux-dey_4.9.bb
	meta-digi-dey/recipes-digi/dey-examples/awsiotsdk-demo_git.bb
	meta-digi-dey/recipes-digi/dey-examples/dey-examples-cryptochip.bb
	meta-digi-dey/recipes-digi/dey-examples/dey-examples-digiapix.bb
	meta-digi-dey/recipes-multimedia/gstreamer/gstreamer1.0-plugins-bad_1.12.%.bbappend
	meta-digi-dey/recipes-multimedia/gstreamer/gstreamer1.0-plugins-base_1.12.%.bbappend
	meta-digi-dey/recipes-multimedia/gstreamer/gstreamer1.0-plugins-good_1.12.%.bbappend
	meta-digi-dey/recipes-multimedia/gstreamer/gstreamer1.0_1.12.%.bbappend
	meta-digi-dey/recipes-multimedia/gstreamer/imx-gst1.0-plugin_4.4.0.bb

+++ meta-digi-dey/recipes-graphics/mesa/mesa_%.bbappend
@@ -1,44 +1,29 @@
 # Copyright (C) 2018 Digi International
-PACKAGECONFIG_remove_imxgpu2d = "egl gles gbm"
+PROVIDES_remove_imxgpu   = "virtual/egl"
+PROVIDES_remove_imxgpu3d = "virtual/libgl virtual/libgles1 virtual/libgles2"
 
-PROVIDES_remove_imxgpu2d = "virtual/libgles1 virtual/libgles2 virtual/libopenvg virtual/egl gbm"
-PROVIDES_remove_imxgpu3d = "virtual/libgl"
-
-# FIXME: mesa should support 'x11-no-tls' option
-python () {
-    overrides = d.getVar("OVERRIDES", True).split(":")
-    if "imxgpu2d" not in overrides:
-        return
-
-    x11flag = d.getVarFlag("PACKAGECONFIG", "x11", False)
-    d.setVarFlag("PACKAGECONFIG", "x11", x11flag.replace("--enable-glx-tls", "--enable-glx"))
-}
-
-# Enable Etnaviv support
-PACKAGECONFIG_append_use-mainline-bsp = " gallium"
-GALLIUMDRIVERS_append_use-mainline-bsp = ",etnaviv,imx"
+PACKAGECONFIG_remove_imxgpu   = "egl gbm"
+PACKAGECONFIG_remove_imxgpu3d = "gles"
 
 BACKEND = \
     "${@bb.utils.contains('DISTRO_FEATURES', 'wayland', 'wayland', \
         bb.utils.contains('DISTRO_FEATURES',     'x11',     'x11', \
                                                              'fb', d), d)}"
-
-# FIXME: Dirty hack to allow use of Vivante GPU libGL binary
 do_install_append_imxgpu3d () {
-    rm -f ${D}${libdir}/libGL.* \
-          ${D}${includedir}/GL/gl.h \
-          ${D}${includedir}/GL/glext.h \
-          ${D}${includedir}/GL/glx.h \
+    rm -f ${D}${includedir}/GL/glx.h \
           ${D}${includedir}/GL/glxext.h
     if [ "${BACKEND}" = "x11" ]; then
         rm -f ${D}${libdir}/pkgconfig/gl.pc
     fi
 }
 
-do_install_append_imxgpu2d () {
-    rm -f ${D}${libdir}/pkgconfig/wayland-egl.pc
+do_install_append_imxgpu () {
+    rm -f ${D}${includedir}/KHR
 }
 
-do_install_append_imxgpu2d () {
-    rm -f ${D}${libdir}/pkgconfig/wayland-egl.pc
+do_install_append () {
+    if [ "${BACKEND}" = "wayland" ]; then
+        rm -f ${D}${libdir}/pkgconfig/wayland-egl.pc
+        rm -f ${D}${libdir}/libwayland-egl*
+    fi
 }
